\name{balance}

\docType{data}

\alias{balance}

\title{Balance Scale Data}

\description{Balance scale data of four distance items from 779 subjects.}

\usage{data(balance)}

\format{Data.frame.}

\source{Jansen ... 2001}

\details{}

\references{Articles} 

\keyword{datasets}
