\name{depmix-internal}

\alias{dens}
\alias{em}
\alias{fb}
\alias{logDens}
\alias{lystig}

\alias{makeDepmix}
\alias{makePriorModel}
\alias{makeResponseModels}
\alias{makeTransModels}

\alias{mlogit}
\alias{multinomial}
\alias{nstates}
\alias{pa2conr}
\alias{viterbi}
\alias{viterbi.fb}

\alias{dens,GLMresponse-method}
\alias{dens,BINOMresponse-method}
\alias{dens,NORMresponse-method}
\alias{dens,MULTINOMresponse-method}
\alias{dens,transInit-method}

\alias{fit,GLMresponse-method}
\alias{fit,BINOMresponse-method}
\alias{fit,NORMresponse-method}w
\alias{fit,MULTINOMresponse-method}
\alias{fit,MVNresponse-method}
\alias{fit,transInit-method}

\alias{logLik,GLMresponse-method}

\alias{predict,GLMresponse-method}
\alias{predict,BINOMresponse-method}
\alias{predict,NORMresponse-method}
\alias{predict,MULTINOMresponse-method}
\alias{predict,transInit-method}

\alias{getpars,GLMresponse-method}
\alias{setpars,GLMresponse-method}

\alias{npar,response-method}

\title{Depmix internal functions}

\description{
  Internal depmix functions, methods and classes.
}

\details{
  These are not to be called by the user.
}

\author{Ingmar Visser & Maarten Speekenbrink}

\keyword{methods}
