\name{formatperc}

\alias{formatperc}

\title{Format percentage for level in printing confidence interval}

\description{See title.}

\usage{formatperc(x,digits)}

\arguments{
	\item{x}{a vector of probabilities to be formatted as percentages.}
	\item{digits}{the number of required digits in the percentages.}
}

\author{Ingmar Visser}