\name{stationary}

\alias{stationary}

\title{Compute the stationary distribution of a transition probability matrix.}

\description{See title.}

\usage{stationary(tpm)}

\arguments{
	\item{tpm}{a transition probability matrix.}
}

\value{A vector with the stationary distribution such that p=tpm*p.}