\name{balance}

\docType{data}

\alias{balance}

\title{Balance Scale Data}

\description{Balance scale data of four distance items from 779 subjects;
participants' ages are included.}

\usage{data(balance)}

\format{Data.frame.}

\source{

	Brenda Jansen (2001),\emph{Development of reasoning on the balance
	scale task: Psychometric assessment of cognitive strategies}.  PhD
	thesis, University of Amsterdam, Department of Psychology.

}

\keyword{datasets}
