\name{balance}

\docType{data}

\alias{balance}

\title{Balance Scale Data}

\description{Balance scale data of four distance items from 779 subjects;
participants' ages are included.}

\usage{data(balance)}

\format{
  
  A data frame with 779 observations on the following variables.

  \describe{
	\item{\code{sex}}{Participants' sex.}
	\item{\code{agedays}}{Age in days.}
	\item{\code{age}}{Age in years.}
	\item{\code{t1}}{Trichotomously scored distance item.}
	\item{\code{t2}}{Trichotomously scored distance item.}
	\item{\code{t3}}{Trichotomously scored distance item.}
	\item{\code{t4}}{Trichotomously scored distance item.}
	\item{\code{d1}}{Dichotomously scored distance item.}
	\item{\code{d2}}{Dichotomously scored distance item.}
	\item{\code{d3}}{Dichotomously scored distance item.}
	\item{\code{d4}}{Dichotomously scored distance item.}
  }
}

\source{

	Brenda Jansen (2001),\emph{Development of reasoning on the balance
	scale task: Psychometric assessment of cognitive strategies}.  PhD
	thesis, University of Amsterdam, Department of Psychology.

}

\examples{
data(balance)
}

\keyword{datasets}
