\name{depmix-internal}

\alias{dens}
\alias{em}
\alias{fb}
\alias{logDens}
\alias{lystig}

\alias{makePriorModel}
\alias{makeResponseModels}
\alias{makeTransModels}

\alias{mlogit}
\alias{multinomial}

\alias{pa2conr}
\alias{viterbi}
\alias{viterbi2}
\alias{viterbi.fb}

\alias{nlin}
\alias{nlin,depmix.fitted-method}
\alias{nlin,mix.fitted-method}

\alias{dens,GLMresponse-method}
\alias{dens,BINOMresponse-method}
\alias{dens,NORMresponse-method}

\alias{dens,GAMMAresponse-method}
\alias{dens,POISSONresponse-method}
\alias{dens,MVNresponse-method}

\alias{dens,MULTINOMresponse-method}
\alias{dens,transInit-method}

\alias{fit,GLMresponse-method}
\alias{fit,BINOMresponse-method}
\alias{fit,NORMresponse-method}

\alias{fit,MULTINOMresponse-method}
\alias{fit,MVNresponse-method}
\alias{fit,transInit-method}

\alias{logLik,GLMresponse-method}

\alias{predict,GLMresponse-method}
\alias{predict,BINOMresponse-method}
\alias{predict,NORMresponse-method}

\alias{predict,MULTINOMresponse-method}
\alias{predict,transInit-method}
\alias{predict,MVNresponse-method}

\alias{getpars,GLMresponse-method}
\alias{setpars,GLMresponse-method}

\alias{getpars,MVNresponse-method}
\alias{setpars,MVNresponse-method}

\alias{npar,response-method}

\alias{MVNresponse,formula-method}


\title{DepmixS4 internal functions}

\description{

  Internal depmix functions, methods and classes.

}

\details{

  These are not to be called by the user.

}

\author{Ingmar Visser}

\keyword{internal}

